Preliminary morning commit
I don't exactly remember what I did in the last hour.
You'll have to look through the diff...

I don't know what kind of barbarian codes in one super file
but then the game is super small so this is completely reasonable.

RIP to whoever is reading this. Hopefully that's not future me.